requirements.txt
click==8.0.4
cycler==0.11.0
fonttools==4.29.1
kiwisolver==1.3.2
matplotlib==3.5.1
numpy==1.22.2
opencv-python==4.5.5.62
packaging==21.3
paho-mqtt==1.6.1
Pillow==9.0.1
pyparsing==3.0.7
python-dateutil==2.8.2
scenedetect==0.5.6.1
six==1.16.0
tqdm==4.63.0
sqlmodel==0.0.6
fastapi==0.75.0
uvicorn==0.17.6
